新增算子 & 代码规范修改



# message auto-generated for no-merge-commit merge:
!266 merge test_a900 into master

新增算子 & 代码规范修改

Created-by: sakuraqqz
Commit-by: 付饶
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
1.新增算子index_fill,fused_cross_entropy_loss_with_max_sum。
2.新增aclnn接口aclnnInplaceMaskedScatter，aclnnInplaceScatterUpdate，aclnnTfScatterAdd。
3.按照issue修改部分代码规范。
## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->
https://gitcode.com/cann/ops-nn/issues/167
https://gitcode.com/cann/ops-nn/issues/158
https://gitcode.com/cann/ops-nn/issues/157
https://gitcode.com/cann/ops-nn/issues/156
## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [x] Bug修复
- [x] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-nn!266

+++ index/scatter_elements_v2/op_kernel/scatter_elements_v2.h
@@ -146,7 +146,7 @@ public:
         uPadParams = {false, 0, 0, static_cast<U>(0)};
     }
 
-    __aicore__ inline void CopyInIndex(int indicesIndex)
+    __aicore__ inline void CopyInIndex(int64_t indicesIndex)
     {
         if constexpr (IS_CAST_INT) {
             DataCopyPadGm2UBImpl(
